consolidate preferences and privacy pages
