--- src/components/Preferences/Preferences.css
@@ -1,22 +1,0 @@
-.Preferences {
-  padding: 0 1rem;
-  margin: 0 0 4rem;
-}
-
-.Preferences .options {
-  padding: 0 0 0 0.5rem;
-  line-height: 1.5;
-}
-
-.Preferences .h2 {
-  background: #376e8c;
-  padding: 0.125rem 0.5rem;
-  color: white;
-  margin: 1rem 0 2rem;
-  border-radius: 2px;
-}
-
-.Preferences input[type="checkbox"] {
-  vertical-align: middle;
-  margin: 0 0.5ex;
-}
